+++ protocol/governance/contracts/modules/core/ElectionModule.sol
@@ -91,17 +91,17 @@ contract ElectionModule is
     }
 
     function setDebtShareSnapshotId(
-        uint snapshotId
+        uint256 snapshotId
     ) external override onlyInPeriod(Council.ElectionPeriod.Nomination) {
         OwnableStorage.onlyOwner();
         _setDebtShareSnapshotId(snapshotId);
     }
 
-    function getDebtShareSnapshotId() external view override returns (uint) {
+    function getDebtShareSnapshotId() external view override returns (uint256) {
         return _getDebtShareSnapshotId();
     }
 
-    function getDebtShare(address user) external view override returns (uint) {
+    function getDebtShare(address user) external view override returns (uint256) {
         return _getDebtShare(user);
     }
 
@@ -111,7 +111,7 @@ contract ElectionModule is
 
     function setCrossChainDebtShareMerkleRoot(
         bytes32 merkleRoot,
-        uint blocknumber
+        uint256 blocknumber
     ) external override onlyInPeriod(Council.ElectionPeriod.Nomination) {
         OwnableStorage.onlyOwner();
         _setCrossChainDebtShareMerkleRoot(merkleRoot, blocknumber);
@@ -127,7 +127,12 @@ contract ElectionModule is
         return _getCrossChainDebtShareMerkleRoot();
     }
 
-    function getCrossChainDebtShareMerkleRootBlockNumber() external view override returns (uint) {
+    function getCrossChainDebtShareMerkleRootBlockNumber()
+        external
+        view
+        override
+        returns (uint256)
+    {
         return _getCrossChainDebtShareMerkleRootBlockNumber();
     }
 
@@ -141,7 +146,7 @@ contract ElectionModule is
         emit CrossChainDebtShareDeclared(user, debtShare);
     }
 
-    function getDeclaredCrossChainDebtShare(address user) external view override returns (uint) {
+    function getDeclaredCrossChainDebtShare(address user) external view override returns (uint256) {
         return _getDeclaredCrossChainDebtShare(user);
     }
 
@@ -159,8 +164,8 @@ contract ElectionModule is
     // Internal
     // ---------------------------------------
 
-    function _sqrt(uint x) internal pure returns (uint y) {
-        uint z = (x + 1) / 2;
+    function _sqrt(uint256 x) internal pure returns (uint256 y) {
+        uint256 z = (x + 1) / 2;
         y = x;
         while (z < y) {
             y = z;
@@ -169,8 +174,8 @@ contract ElectionModule is
     }
 
     /// @dev Overrides the user's voting power by combining local chain debt share with debt shares in other chains, quadratically filtered
-    function _getVotePower(address user) internal view virtual override returns (uint) {
-        uint votePower = _getDebtShare(user) + _getDeclaredCrossChainDebtShare(user);
+    function _getVotePower(address user) internal view virtual override returns (uint256) {
+        uint256 votePower = _getDebtShare(user) + _getDeclaredCrossChainDebtShare(user);
 
         return _sqrt(votePower);
     }
